Make recordAsRecord public
