--- src/internal/empowerRecord.js
@@ -1,8 +1,0 @@
-// @flow
-
-export default (fn: Function, record: *, returnSelf: boolean): * => {
-    let objectResult = fn(record.toObject());
-    return returnSelf
-        ? record.__proto__.constructor(objectResult)
-        : objectResult;
-};
