+++ src/util/recordAsObject.js
@@ -1,0 +1,8 @@
+// @flow
+
+export default (updater: Function, record: *, returnRecord: boolean = true): * => {
+    let result = updater(record.toObject());
+    return returnRecord
+        ? record.__proto__.constructor(result)
+        : result;
+};
